split tests in ci
