+++ .github/workflows/python.yaml
@@ -288,31 +288,32 @@ jobs:
             --timeout=1500 \
             -W ignore::PendingDeprecationWarning \
             --cov-config=setup.cfg --cov-report=xml --cov=xorbits \
+            -k "not test_execution_with_process_exit_message"
             xorbits
           # workaround: this case will hang, run it separately.
           pytest --timeout=1500 \
             -W ignore::PendingDeprecationWarning \
             --cov-config=setup.cfg --cov-report=xml --cov=xorbits \
-            xorbits/core/tests/test_execution_exit.py
+            -k "test_execution_with_process_exit_message" \
+            xorbits/core/tests/test_execution.py
         elif [[ "$MODULE" == "xorbits/pandas" ]]; then
           pytest --timeout=1500 \
             -W ignore::PendingDeprecationWarning \
             --cov-config=setup.cfg --cov-report=xml \
             --cov=xorbits xorbits/pandas
         elif [[ "$MODULE" == "xorbits/numpy" ]]; then
-          pytest --timeout=1500 --ignore xorbits/numpy/numpy_adapters \
+          pytest --timeout=1500 \
             -W ignore::PendingDeprecationWarning \
-            --cov-config=setup.cfg --cov-report=xml \
-            --cov=xorbits xorbits/numpy
+            --cov-config=setup.cfg --cov-report=xml --cov=xorbits \
+            -k "not test_numpy_fallback" \
+            xorbits/numpy
           # workaround: this case will hang, run it separately.
           pytest --timeout=1500 \
             -W ignore::PendingDeprecationWarning \
             --cov-config=setup.cfg --cov-report=xml \
-            --cov=xorbits xorbits/numpy/numpy_adapters/tests/test_numpy_adapters.py
-          pytest --timeout=1500 \
-            -W ignore::PendingDeprecationWarning \
-            --cov-config=setup.cfg --cov-report=xml \
-            --cov=xorbits xorbits/numpy/numpy_adapters/tests/test_numpy_fallback.py
+            --cov=xorbits \
+            -k "test_numpy_fallback" \
+            xorbits/numpy/numpy_adapters/tests/test_numpy_adapters.py
         elif [[ "$MODULE" == "mars-core" ]]; then
           pytest --forked --log-level=DEBUG --ignore xorbits/_mars/dataframe --ignore xorbits/_mars/tensor \
             --ignore xorbits/_mars/learn --ignore xorbits/_mars/remote \
